update
git-svn-id: http://svn.tudos.org/repos/tudos/trunk@137 a704ac0b-3a55-4d43-a2a9-7be6f07c34fb

+++ kernel/fiasco/src/kern/jdb_lines.cpp
@@ -5,7 +5,7 @@ INTERFACE:
 class Jdb_lines_line
 {
 public:
-  unsigned int   addr;
+  unsigned long  addr;
   unsigned short line;
 } __attribute__((packed));
